fix gemfile version, make string improvements, fix JRuby use
Gemfile now specifies a minimum version for ffi.
Strings are now read in by first scanning the length and then loading in
as a byte array, rather than reading in each individual unit and then
packing them.  I'm not sure which method is functionally faster, but
neither is extremely slow, and this way is clearer and easier to read.
JRuby FFI breaks on read_uint16 for some reason.  Only get_uint16(0)
works.  This is probably because the jruby ffi gem is a completely
different implementation